Fully working collision
Food spawns in new location when collision happens, and game freezes after hitting snake body and the wall. Also fixed a movement glitch where if two keys are tapped in rapid succession, it could freeze the snake. Next is using text to create a more user friendly experience, as well as adding score.